refactor: Standardize all table ID columns to auto-increment and use UUID for leagues
Schema Changes:
- sports.id: INT → BIGSERIAL (auto-increment 64-bit)
- venues.id: INT → BIGSERIAL (auto-increment 64-bit)
- leagues.id: INT → UUID with DEFAULT gen_random_uuid()
- leagues.sport_id: INT → BIGINT (foreign key reference)
- leagues.venue_id: INT → BIGINT (foreign key reference)
- game_occurrences.id: SERIAL → BIGSERIAL (auto-increment 64-bit)
- game_occurrences.league_id: INT → UUID (foreign key reference)
- leagues_drafts.id: INT → BIGSERIAL (auto-increment 64-bit)
- leagues_drafts: Fixed UNIQUE constraint to composite (org_id, type)
- Added name column to leagues_drafts for user-defined draft/template names

Backend Updates:
- Sport.ID: int → int64
- Venue.ID: int → int64
- League.ID: int → *string (UUID)
- League.SportID/VenueID: *int → *int64
- Added UUID-based league methods: GetLeagueByUUID, ApproveLeagueByUUID, RejectLeagueByUUID
- Updated handlers to work with UUID strings instead of int conversion
- Fixed repository to extract UUID from Supabase responses

Frontend Updates:
- Fixed SupabaseContext TypeScript error by converting to async/await pattern
- SupabaseProvider now skips initialization when no session (auth pages)
- Reduced dependency injection in Supabase initialization

Fixes:
- Removed invalid game_occurrences and supplemental_requests from league insert
- UpdateLastLogin no longer uses invalid __increment syntax
- All table IDs now properly auto-increment where appropriate

+++ backend/internal/sports/repository.go
@@ -92,7 +92,7 @@ func (r *Repository) Create(ctx context.Context, name string) (*Sport, error) {
 	// Extract ID from result if available
 	if len(result) > 0 && result[0]["id"] != nil {
 		if id, ok := result[0]["id"].(float64); ok {
-			sport.ID = int(id)
+			sport.ID = int64(id)
 		}
 	}
